+++ git_bbb/__init__.py
@@ -1,5 +1,7 @@
 from prompt_toolkit import Application
 from prompt_toolkit.layout import Layout
+from prompt_toolkit.styles.pygments import style_from_pygments_cls
+from pygments.styles import get_style_by_name
 
 from .git_plumbing import default_ignore_revs
 from .browser import Browser, browse_blame_briskly
@@ -28,11 +30,14 @@ def run(path, rev, ignore_revs_file):
         browser, undo_redo_stack, ignore_revs_file
     )
 
+    # TODO: make this configurable
+    pygments_style = "monokai"
     app = Application(
         full_screen=True,
         layout=layout,
         key_bindings=key_bindings,
         mouse_support=True,
+        style=style_from_pygments_cls(get_style_by_name(pygments_style)),
     )
 
     app.run()